Add is_strange functionality
Add method "is_strange" to the class Passcard in models.py, this
method checks if the passcard's owner has made at least one long visit
to the storage and returns True if he has, False otherwise. The method
takes an optional argument minutes with a default value of 60, that
stands for defining visit duration which should be considered strange.
Also add 'is_strange' information to visit_info dictionary in
non_closed_visits list to display it on the page with
storage information.
It's important to have this functionality to identify any
suspicious activity in the storage. It is displaying now whether person
in the storage is suspicious or not.